разделил модуль компаненты из content.js на отдельные модули
