+++ src/components/event.js
@@ -1,22 +1,3 @@
-export const createTripContainerTemplate = () => {
-  return (`<ul class="trip-days"></ul>`);
-};
-
-export const createTripDayContainerTemplate = () => {
-  return (
-    `<li class="trip-days__item  day">
-        <div class="day__info">
-          <span class="day__counter">1</span>
-          <time class="day__date" datetime="2019-03-18">MAR 18</time>
-        </div>
-     </li>`
-  );
-};
-
-export const createTripEventTemplate = () => {
-  return (`<ul class="trip-events__list"></ul>`);
-};
-
 export const createEventItemTemplate = () => {
   return (
     `<li class="trip-events__item">
