+++ src/components/trip-container.js
@@ -1,0 +1,3 @@
+export const createTripContainerTemplate = () => {
+  return (`<ul class="trip-days"></ul>`);
+};
