+++ src/components/trip-events.js
@@ -1,0 +1,3 @@
+export const createTripEventTemplate = () => {
+  return (`<ul class="trip-events__list"></ul>`);
+};
